requirements.txt
anyio
certifi==2026.1.4
charset-normalizer==3.4.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
python-telegram-bot==21.6
requests==2.32.4
typing_extensions
urllib3
pymongo
dnspython
redis==5.0.3